+++ operators/__init__.py
@@ -4,10 +4,13 @@ from . import (
     tracking,
     cleanup_tracks,
     cleanup_operator,
-    setup_defaults,
     test_setup_defaults,
     test_variants,
-    error_value,
+)
+
+from ..helpers import (
+    track_default_settings,
+    error_value_operator,
 )
 
 operator_classes = (
@@ -15,8 +18,8 @@ operator_classes = (
     *tracking.operator_classes,
     *cleanup_tracks.operator_classes,
     *cleanup_operator.operator_classes,
-    *setup_defaults.operator_classes,
     *test_setup_defaults.operator_classes,
     *test_variants.operator_classes,
-    *error_value.operator_classes,
+    *track_default_settings.operator_classes,
+    *error_value_operator.operator_classes,
 )
